// src/app/core/models/meeting.ts
import { Committee } from './committee';
import { MeetingParticipant } from './meeting-participant';
import { TimeZone } from './time-zone';
import { BaseModel } from './baseModel';
import { User } from './user';
import { MeetingAgenda } from './meeting-agenda';
import { Attachment } from './attachment';
import { Proposal } from './proposal';
import { UserComment } from './user-comment';
import { MeetingVote } from './meeting-vote';
import { MinOfMeeting } from './min-of-meeting';
import { Task } from './task';
import { MeetingRecommendation } from './MeetingRecommendation';
export class Meeting extends BaseModel {
	id: number;
	meeting_title_en: string;
	meeting_title_ar: string;
	meeting_venue_ar: string;
	meeting_venue_en: string;
	meeting_type_id: number;
	meeting_mom_template_id: number;
	meeting_status_id: number;
	meeting_status_name_ar: string;
	meeting_status_name_en: string;
	meeting_description_ar: string;
	meeting_description_en: string;
	time_zone_id: number;
	meeting_note_ar: string;
	meeting_note_en: string;
	organization_id: number;
	meeting_reminders: Array<number>;
	meeting_schedule_to: any;
	meeting_schedule_from: any;
	meeting_schedule_from_date: any;
	meeting_schedule_to_date: any;
	meeting_schedule_from_time: any;
	meeting_schedule_to_time: any;
	committee_id: number;
	meeting_code: string;
	meeting_participants: Array<MeetingParticipant>;
	guests: Array<any>;
	meeting_organisers: Array<User>;
	meeting_agendas: Array<MeetingAgenda>;
	meeting_attachments: Array<Attachment>;
	time_zone: TimeZone;
	proposal = new Proposal();
	participant_meeting_attendance_status_id: number;
	meeting_user_comment = new UserComment();
	location_lat: number;
	location_long: number;
	meetingMemberIds: Array<number>;
	proposal_id: number;
	can_edit_meeting: boolean;
	meeting_votes: Array<MeetingVote>;
	meeting_committee = new Committee();
	can_attend: boolean;
	meeting: boolean;
	committee_name_ar: string;
	committee_name_en: string;
	is_started: boolean;
	grouped_meeting_participants: Array<any>;
	is_mom_sent: boolean;
	current_agenda: MeetingAgenda;
	meetingAgendas?: Array<MeetingAgenda> = [];
	can_start_vote: boolean;
	can_add_vote: boolean;
	show_send_sign_button: boolean;
	is_signature_sent: boolean;
	can_sign: boolean;
	participant_meeting_attendance_status_name_ar: string;
	participant_meeting_attendance_status_name_en: string;
	can_change_meeting_status: boolean;
	meeting_day_name_ar: string;
	meeting_month: string;
	meeting_day: string;
	meeting_month_ar: string;
	meeting_day_name: string;
	meeting_tasks: Array<Task>;
	zoom_join_url: string;
	zoom_start_url: string;
	is_zoom_enabled: boolean;
	zoom_meeting_id: string;
	is_selected: boolean;
	last_message_text: string;
	last_message_date: any;
	chat_room_id: number;
	is_microsoft_teams_enabled: boolean;
	microsoft_teams_meeting_id: number;
	online_meeting_start_url: string;
	is_online_meeting_enable: boolean;
	online_meeting_join_url: string;
	online_meeting_app_id: number;
	online_meeting_type_id: number;
	meeting_version_id: number;
	is_changed_publish: boolean;
	is_published: boolean;
	related_meeting_id: number;
	is_attachment_hidden: boolean;
	is_agenda_hidden: boolean;
	online_configuration_id: number;
	is_online_configuration_active: boolean;
	participantStatistics: any;
	can_view_meeting_statistic: boolean;
	meeting_attendance_percentage: number;
	show_attendance_percentage_warning: number;
	can_change_attendance: boolean;
	show_timer: boolean;
	is_mom_pdf: boolean;
	mom_pdf_url: string;
	mom_pdf_file_name: string;
	meeting_stakeholders_percentage: number;
	meetingRecommendations: Array<MeetingRecommendation>;
	can_view_recommendation: boolean;
	can_edit_recommendation_meeting: boolean;
}